refector, todo : chatMessage 엔티티 수정, controller, dto, service에 todo 추가 #5

+++ src/main/java/com/ktcloud/daangn/chat/dto/ChatRoomReadResponseDto.java
@@ -3,6 +3,6 @@ package com.ktcloud.daangn.chat.dto;
 public record ChatRoomReadResponseDto(
         Long roomId,
         Long memberId,
-        int readMessageCount
+        Long readMessageCount
 ) {
 }
